requirements.txt
flask==2.3.2
flask-limiter==2.8.1
python-dotenv==1.0.0
dnspython==2.3.0
python-whois==0.8.0
